backend_legacy/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
appwrite==14.1.0
pydantic[email]==2.10.5
cryptography==44.0.0
httpx==0.28.1
aiosmtplib==3.0.2
python-multipart==0.0.20
pydantic-settings
cloudinary==1.41.0